Expose the client and dashboard abstracton.
This is basically a giftware commit of the entire project.

Parts of this are taken from the performanceplatform-notifier with the intent of sharing this client.

Because the notifiers client is quite implementation specific we can only share upto the point of querist.

In order to use Querist we expose it on the main index alongside the Dashboard abstraction.