Removed Querist. It wasn't doing much for us.
We have a problem with testing at the moment.

We can really only validate these tests against nodejs as mocking primative modules is hard to do in browserify at present.

We could use sinon.fakeServer to mock the http but that doesn't work at present.

Once sinon 2.0 is available we'll refactor to use that.

Right now our tests are really only going to run in the pure nodes mode and not in node and karma.